--- src/pykeen/models/nmodel/__init__.py
@@ -1,6 +1,0 @@
-# -*- coding: utf-8 -*-
-
-"""New-style models."""
-
-from .base import ERModel, _NewAbstractModel  # noqa
-from .representation import EmbeddingSpecification  # noqa
