+++ tox.ini
@@ -130,7 +130,8 @@ commands = mypy --ignore-missing-imports \
     src/pykeen/nn \
     src/pykeen/models/unimodal/ \
     src/pykeen/models/multimodal/ \
-    src/pykeen/models/nmodel/ \
+    src/pykeen/models/nbase.py \
+    src/pykeen/models/nemb.py \
     src/pykeen/models/base.py \
     src/pykeen/models/predict.py \
     src/pykeen/models/__init__.py \
